Beginning the refactor
