+++ lib/templates.ts
@@ -1,7 +1,6 @@
 import { formCache } from "./cache/formCache";
 import { prisma, prismaErrors, Prisma } from "@gcforms/database";
 import {
-  PublicFormRecord,
   FormRecord,
   FormProperties,
   DeliveryOption,
@@ -9,7 +8,7 @@ import {
   ClosedDetails,
 } from "@lib/types";
 
-import { authorization, getAbility } from "./privileges";
+import { authorization } from "./privileges";
 import { AuditLogAccessDeniedDetails, AuditLogDetails, AuditLogEvent, logEvent } from "./auditLogs";
 import { logMessage } from "@lib/logger";
 import { unprocessedSubmissions, deleteDraftFormResponses } from "./vault";
@@ -23,95 +22,9 @@ import { validateTemplate } from "@lib/utils/form-builder/validate";
 import { dateHasPast } from "@lib/utils";
 import { validateTemplateSize } from "@lib/utils/validateTemplateSize";
 import { NotificationsInterval } from "@gcforms/types";
-import { checkOne } from "@lib/cache/flags";
+
 import { checkForBetaComponentsAsync } from "./validation/betaCheck";
 import { invalidateTemplateEditLockUserCountCache } from "./editLocks";
-
-const checkFlag = async (flag: string) => {
-  return (await Promise.all([checkOne(flag), authorization.canAccessBetaComponents(flag)])).reduce(
-    (prev, curr) => prev || curr
-  );
-};
-
-// ******************************************
-// Internal Module Functions
-// ******************************************
-const _parseTemplate = (template: {
-  id: string;
-  created_at?: Date;
-  updated_at?: Date;
-  name: string;
-  jsonConfig: Prisma.JsonValue;
-  isPublished: boolean;
-  deliveryOption: {
-    emailAddress: string;
-    emailSubjectEn: string | null;
-    emailSubjectFr: string | null;
-  } | null;
-  securityAttribute: string;
-  formPurpose: string;
-  publishReason: string;
-  publishFormType: string;
-  publishDesc: string;
-  closingDate?: Date | null;
-  closedDetails?: Prisma.JsonValue | null;
-  saveAndResume: boolean;
-  notificationsInterval?: number | null;
-  ttl?: Date | null;
-}): FormRecord => {
-  return {
-    id: template.id,
-    ...(template.created_at && {
-      createdAt: template.created_at?.toString(),
-    }),
-    ...(template.updated_at && {
-      updatedAt: template.updated_at.toString(),
-    }),
-    name: template.name,
-    form: template.jsonConfig as FormProperties,
-    isPublished: template.isPublished,
-    ...(template.deliveryOption && {
-      deliveryOption: {
-        emailAddress: template.deliveryOption.emailAddress,
-        ...(template.deliveryOption.emailSubjectEn && {
-          emailSubjectEn: template.deliveryOption.emailSubjectEn,
-        }),
-        ...(template.deliveryOption.emailSubjectFr && {
-          emailSubjectFr: template.deliveryOption.emailSubjectFr,
-        }),
-      },
-    }),
-    formPurpose: template.formPurpose,
-    publishReason: template.publishReason,
-    publishFormType: template.publishFormType,
-    publishDesc: template.publishDesc,
-    securityAttribute: template.securityAttribute as SecurityAttribute,
-    ...(template.closingDate && {
-      closingDate: template.closingDate.toString(),
-    }),
-    closedDetails: template.closedDetails as ClosedDetails,
-    saveAndResume: template.saveAndResume,
-    notificationsInterval: template.notificationsInterval as NotificationsInterval,
-    ...(template.ttl && { ttl: template.ttl }),
-  };
-};
-
-// ******************************************
-// Exportable Module Functions
-// ******************************************
-
-export type CreateTemplateCommand = {
-  userID: string;
-  formConfig: FormProperties;
-  name?: string;
-  deliveryOption?: DeliveryOption;
-  securityAttribute?: SecurityAttribute;
-  formPurpose?: string;
-  publishReason?: string;
-  publishFormType?: string;
-  publishDesc?: string;
-  notificationsInterval?: NotificationsInterval;
-};
 
 export type UpdateTemplateCommand = {
   formID: string;
@@ -125,320 +38,6 @@ export type UpdateTemplateCommand = {
   publishDesc?: string;
   notificationsInterval?: NotificationsInterval;
 };
-
-export class InvalidFormConfigError extends Error {
-  constructor(message?: string) {
-    super(message ?? "InvalidFormConfigError");
-    Object.setPrototypeOf(this, InvalidFormConfigError.prototype);
-  }
-}
-
-export class TemplateAlreadyPublishedError extends Error {
-  constructor(message?: string) {
-    super(message ?? "TemplateAlreadyPublishedError");
-    Object.setPrototypeOf(this, TemplateAlreadyPublishedError.prototype);
-  }
-}
-
-export class TemplateHasUnprocessedSubmissions extends Error {
-  constructor(message?: string) {
-    super(message ?? "TemplateHasUnprocessedSubmissions");
-    Object.setPrototypeOf(this, TemplateHasUnprocessedSubmissions.prototype);
-  }
-}
-
-/**
- * Creates a Form Template record
- * @param config Form Template configuration
- * @returns Form Record or null if creation was not sucessfull.
- */
-export async function createTemplate(command: CreateTemplateCommand): Promise<FormRecord | null> {
-  const { user } = await authorization.canCreateForm().catch((e) => {
-    logEvent(
-      e.user.id,
-      { type: "Form" },
-      "AccessDenied",
-      AuditLogAccessDeniedDetails.AccessDenied_AttemptToCreateForm
-    );
-    throw e;
-  });
-
-  await checkForBetaComponentsAsync(command.formConfig.elements, checkFlag);
-
-  const validationResult = validateTemplate(command.formConfig);
-
-  if (!validationResult.valid) {
-    logMessage.warn(
-      `[templates][createTemplate] Form config is invalid.\nReasons: ${JSON.stringify(
-        validationResult.errors
-      )}.\nConfig: ${JSON.stringify(command.formConfig)}`
-    );
-    throw new InvalidFormConfigError();
-  }
-
-  const isValid = validateTemplateSize(JSON.stringify(command.formConfig));
-
-  if (!isValid) {
-    logMessage.warn(
-      `[templates][createTemplate] Template size exceeds the limit.\nConfig: ${JSON.stringify(
-        command.formConfig
-      )}`
-    );
-    throw new InvalidFormConfigError();
-  }
-
-  const createdTemplate = await prisma.template
-    .create({
-      data: {
-        jsonConfig: command.formConfig as Prisma.JsonObject,
-        ...(command.name && {
-          name: command.name,
-        }),
-        ...(command.deliveryOption && {
-          deliveryOption: {
-            create: {
-              emailAddress: command.deliveryOption.emailAddress,
-              emailSubjectEn: command.deliveryOption.emailSubjectEn,
-              emailSubjectFr: command.deliveryOption.emailSubjectFr,
-            },
-          },
-        }),
-        ...(command.securityAttribute && {
-          securityAttribute: command.securityAttribute as string,
-        }),
-        users: {
-          connect: { id: command.userID },
-        },
-        ...(command.formPurpose && { formPurpose: command.formPurpose }),
-        ...(command.notificationsInterval !== undefined && {
-          notificationsInterval: command.notificationsInterval,
-        }),
-      },
-      select: {
-        id: true,
-        created_at: true,
-        updated_at: true,
-        name: true,
-        jsonConfig: true,
-        isPublished: true,
-        deliveryOption: true,
-        securityAttribute: true,
-        formPurpose: true,
-        publishReason: true,
-        publishFormType: true,
-        publishDesc: true,
-        saveAndResume: true,
-        notificationsInterval: true,
-      },
-    })
-    .catch((e) => prismaErrors(e, null));
-
-  if (createdTemplate === null) return null;
-
-  logEvent(user.id, { type: "Form", id: createdTemplate?.id }, "CreateForm");
-
-  return _parseTemplate(createdTemplate);
-}
-
-/**
- * Get all form templates. Must has Manage All Forms privilege.
- * @returns An array of Form Records
- */
-export async function getAllTemplates(options?: {
-  requestedWhere?: Prisma.TemplateWhereInput;
-  sortByDateUpdated?: "asc" | "desc";
-}): Promise<Array<FormRecord>> {
-  try {
-    const { requestedWhere, sortByDateUpdated } = options ?? {};
-    // Can a user view any Template
-    const { user } = await authorization.canViewAllForms().catch((e) => {
-      logEvent(
-        e.user.id,
-        { type: "Form" },
-        "AccessDenied",
-        AuditLogAccessDeniedDetails.AccessDenied_AttempttoAccessAllSystemForms
-      );
-      throw e;
-    });
-
-    const templates = await prisma.template
-      .findMany({
-        where: {
-          ...(requestedWhere && requestedWhere),
-        },
-        select: {
-          id: true,
-          created_at: true,
-          updated_at: true,
-          name: true,
-          jsonConfig: true,
-          isPublished: true,
-          deliveryOption: true,
-          securityAttribute: true,
-          formPurpose: true,
-          publishReason: true,
-          publishFormType: true,
-          publishDesc: true,
-          saveAndResume: true,
-          notificationsInterval: true,
-        },
-        ...(sortByDateUpdated && {
-          orderBy: {
-            updated_at: sortByDateUpdated,
-          },
-        }),
-      })
-      .catch((e) => prismaErrors(e, []));
-
-    // Only log the event if templates are found
-    if (templates.length > 0)
-      logEvent(user.id, { type: "Form" }, "ReadForm", AuditLogDetails.AccessedAllSystemForms);
-
-    return templates.map((template) => _parseTemplate(template));
-  } catch (e) {
-    logMessage.error(e);
-    return [];
-  }
-}
-
-export type TemplateOptions = {
-  sortByDateUpdated?: "asc" | "desc";
-  requestedWhere?: Prisma.TemplateWhereInput;
-};
-
-/**
- * Get all form templates for the User calling the function.
- * @returns An array of Form Records
- */
-export async function getAllTemplatesForUser(
-  options?: TemplateOptions
-): Promise<Array<FormRecord>> {
-  try {
-    const ability = await getAbility();
-
-    const { sortByDateUpdated, requestedWhere } = options ?? {};
-    const templates = await prisma.template
-      .findMany({
-        where: {
-          ...(requestedWhere && requestedWhere),
-          users: {
-            some: {
-              id: ability.user.id,
-            },
-          },
-        },
-        select: {
-          id: true,
-          created_at: true,
-          updated_at: true,
-          ttl: true,
-          name: true,
-          jsonConfig: true,
-          isPublished: true,
-          deliveryOption: true,
-          securityAttribute: true,
-          formPurpose: true,
-          publishReason: true,
-          publishFormType: true,
-          publishDesc: true,
-          saveAndResume: true,
-          notificationsInterval: true,
-        },
-        ...(sortByDateUpdated && {
-          orderBy: {
-            updated_at: sortByDateUpdated,
-          },
-        }),
-      })
-      .catch((e) => prismaErrors(e, []));
-
-    // Only log the event if templates are found
-    if (templates.length > 0)
-      logEvent(ability.user.id, { type: "Form" }, "ReadForm", AuditLogDetails.AccessedForms, {
-        formList: templates.map((template) => template.id).toString(),
-      });
-
-    return templates.map((template) => _parseTemplate(template));
-  } catch (e) {
-    logMessage.error(e);
-    return [];
-  }
-}
-
-/**
- * Get a form template by ID (only includes public information but does not require any permission)
- * @param formID ID of form template
- * @returns PublicFormRecord
- */
-export async function getPublicTemplateByID(formID: string): Promise<PublicFormRecord | null> {
-  try {
-    if (formCache.cacheAvailable) {
-      // This value will always be the latest if it exists because
-      // the cache is invalidated on change of a template
-      const cachedValue = await formCache.check(formID);
-      if (cachedValue) {
-        return cachedValue;
-      }
-    }
-
-    const template = await prisma.template
-      .findUnique({
-        where: {
-          id: formID,
-        },
-        select: {
-          id: true,
-          created_at: true,
-          updated_at: true,
-          name: true,
-          jsonConfig: true,
-          isPublished: true,
-          deliveryOption: true,
-          securityAttribute: true,
-          formPurpose: true,
-          publishReason: true,
-          publishFormType: true,
-          publishDesc: true,
-          closingDate: true,
-          closedDetails: true,
-          saveAndResume: true,
-          ttl: true,
-          notificationsInterval: true,
-        },
-      })
-      .catch((e) => prismaErrors(e, null));
-
-    // Short circuit the public record filtering if no form record is found or the form is marked as deleted (ttl != null)
-    if (!template || template.ttl) return null;
-
-    const parsedTemplate = _parseTemplate(template);
-    const publicFormRecord = onlyIncludePublicProperties(parsedTemplate);
-
-    if (formCache.cacheAvailable) formCache.set(formID, publicFormRecord);
-
-    return publicFormRecord;
-  } catch (e) {
-    logMessage.error(e);
-    return null;
-  }
-}
-
-export async function getTemplatePublishedStatus(formID: string): Promise<boolean | null> {
-  const template = await prisma.template
-    .findUnique({
-      where: {
-        id: formID,
-      },
-      select: {
-        isPublished: true,
-      },
-    })
-    .catch((e) => prismaErrors(e, null));
-
-  if (!template) return null;
-
-  return template.isPublished;
-}
 
 /**
  * Get a form template by ID (includes full template information but requires view permission)
@@ -1566,27 +1165,6 @@ export async function restoreTemplate(formID: string): Promise<FormRecord | null
 // Remove and replace this utility with new authorization object in code
 export const checkUserHasTemplateOwnership = async (formID: string) => {
   await authorization.canEditForm(formID);
-};
-
-/*
- * Extract only the public properties from a form record.
- * The public properties are the ones that are needed to display the form
- * to unauthenticated users. (e.g. when filling out a form)
- * Also sets some of the default values for properties that are not set.
- * @param template A Form Record, containing all the properties
- * @returns a Public Form Record, with only the public properties
- */
-export const onlyIncludePublicProperties = (template: FormRecord): PublicFormRecord => {
-  return {
-    id: template.id,
-    updatedAt: template.updatedAt,
-    closingDate: template.closingDate,
-    closedDetails: template.closedDetails,
-    form: template.form,
-    isPublished: template.isPublished,
-    securityAttribute: template.securityAttribute,
-    saveAndResume: template.saveAndResume,
-  };
 };
 
 export const updateClosedData = async (
